Chat sessions, and fixing build script.
Chat is now sessioned, although there is currently only one session for
the entirety of the server's lifetime, and no attempt is made at storing
the session. Chat is prerendered using Coffeecup, and there's even a
little fanciness to ensure that messages made in between the page
loading and the socket logging in are still displayed on the screen. I
think this means that some messages are capable of being displayed twice
(if they arrive from the 'chat' message AND from the 'logged in' message
but it's unlikely to be an issue unless someone has long loading times.
Even then, no information is lost. It might be worth looking into later.

The advanced optimisations of Google's Closure Compiler were causing
issues. The Cakefile now only applies the default optimisations, which
appears to have sped up the compilation process as well. Watching the
project now produces alerts when recompiling, though it appears to be
setting itself off, as it prints three times for every change.