Adding direct user messaging.
Chatters can now 'whisper' to one another. This will send their messages
to only the selected user. Note that if the user you are trying to
whisper to logs off while you are writing a message, the selection will
swap back to public. This is probably not a good thing.

Chat messages now also avoid reproducing a username if it is directly
above the written message.