Adding map functionality to client.
The map is now communicated from the server to the client. The browser
then renders this map onto the canvas. At the moment the tiles have a
placeholder 'color' property to show the statefulness of the server
communication.

The client can drag the map around. It will not allow the map to be
dragged beyond its vertical edges, and the map loops forever
horizontally.

The client can also zoom in and out of the map, within arbitrary limits.
This functionality is relatively experimental: it only zooms on the top
left corner. In the future it would be nice to zoom on the mouse.
Apparantly the mousewheel event is buggy in some versions of IE and not
supported at all in Firefox. Not tested outside of Chrome.