Setting text y position ~correctly.
It's not clear that the value is really what it should be, but the
result looks fairly well aligned.

This commit also removes the faulty selected tile rendering. A lot of
the canvas interaction in map.js should probably be moved into draw.js,
which will make rendering this selection over the top of the map much
easier.